+++ BeeKit/Model/Goal.swift
@@ -191,126 +191,3 @@ public class Goal: NSManagedObject {
         lastModifiedLocal = Date()
     }
 }
-
-@objc(DueByTableValueTransformer)
-public class DueByTableValueTransformer: ValueTransformer {
-    public override class func transformedValueClass() -> AnyClass {
-        return NSData.self
-    }
-    
-    public override func transformedValue(_ value: Any?) -> Any? {
-        guard let dueByTable = value as? DueByTable else { return nil }
-        do {
-            return try JSONEncoder().encode(dueByTable)
-        } catch {
-            print("Error encoding due by table: \(error)")
-        }
-        return nil
-    }
-
-    public override func reverseTransformedValue(_ value: Any?) -> Any? {
-        guard let data = (value as? Data) else { return nil }
-        do {
-            return try JSONDecoder().decode(DueByTable.self, from: data)
-        } catch {
-            print("Error decoding dictionary: \(error)")
-            return nil
-        }
-    }
-    
-    public static var name: NSValueTransformerName {
-        .init(rawValue: String(describing: DueByTableValueTransformer.self))
-    }
-    
-    public static func register() {
-        ValueTransformer.setValueTransformer(DueByTableValueTransformer(),
-                                             forName: DueByTableValueTransformer.name)
-    }
-}
-
-@objc(DueByTable)
-public class DueByTable: NSObject, NSSecureCoding, Codable {
-    public private(set) var entries: [String: BeeminderDueByEntry]
-    
-    private enum Key: String {
-        case entries
-    }
-    
-    public init(entries: [String: BeeminderDueByEntry]) {
-        self.entries = entries
-    }
-    
-    public static var supportsSecureCoding: Bool { true }
-    
-    public func encode(with coder: NSCoder) {
-        coder.encode(entries, forKey: Key.entries.rawValue)
-    }
-    
-    public required convenience init?(coder: NSCoder) {
-        let entries = coder.decodeObject(forKey: Key.entries.rawValue) as? [String: BeeminderDueByEntry] ?? [:]
-        
-        self.init(entries: entries)
-    }
-    
-    convenience init(dueByJson: JSON?) {
-        var entries: [String : BeeminderDueByEntry] {
-            dueByJson?.dictionary?.compactMapValues(BeeminderDueByEntry.init)
-            ?? dueByJson?.dictionary?.mapValues(BeeminderDueByEntry.init)
-            ?? [:]
-        }
-        self.init(entries: entries)
-    }
-}
-
-public class BeeminderDueByEntry: NSObject, NSSecureCoding, Codable {
-    public let total: Double
-    public let delta: Double
-    public let formatted_total_for_beedroid: String
-    public let formatted_delta_for_beedroid: String
-    
-    public init(json: JSON) {
-        self.delta = json["delta"].doubleValue
-        self.total = json["total"].doubleValue
-        
-        self.formatted_delta_for_beedroid = json["formatted_delta_for_beedroid"].stringValue
-        self.formatted_total_for_beedroid = json["formatted_total_for_beedroid"].stringValue
-    }
-    
-    init(total: Double, delta: Double, formatted_total_for_beedroid: String, formatted_delta_for_beedroid: String) {
-        self.total = total
-        self.delta = delta
-        self.formatted_total_for_beedroid = formatted_total_for_beedroid
-        self.formatted_delta_for_beedroid = formatted_delta_for_beedroid
-    }
-    
-    public static var supportsSecureCoding: Bool { true }
-
-    private enum Key: String {
-        case total
-        case delta
-        case formattedTotalForBeedroid
-        case formattedDeltaForBeedroid
-    }
-    
-    public func encode(with coder: NSCoder) {
-        coder.encode(total, forKey: Key.total.rawValue)
-        coder.encode(delta, forKey: Key.delta.rawValue)
-        coder.encode(formatted_total_for_beedroid, forKey: Key.formattedTotalForBeedroid.rawValue)
-        coder.encode(formatted_delta_for_beedroid, forKey: Key.formattedDeltaForBeedroid.rawValue)
-    }
-    
-    public required convenience init?(coder: NSCoder) {
-        let total = coder.decodeDouble(forKey: Key.total.rawValue)
-        let delta = coder.decodeDouble(forKey: Key.delta.rawValue)
-        
-        guard
-            let formatted_total_for_beedroid = coder.decodeObject(of: NSString.self, forKey: Key.formattedTotalForBeedroid.rawValue) as? String,
-            let formatted_delta_for_beedroid = coder.decodeObject(of: NSString.self, forKey: Key.formattedDeltaForBeedroid.rawValue) as? String
-        else { return nil }
-        
-        self.init(total: total,
-                  delta: delta,
-                  formatted_total_for_beedroid: formatted_total_for_beedroid,
-                  formatted_delta_for_beedroid: formatted_delta_for_beedroid)
-    }
-}
